feat: use same signature as compute ft_fma (#220)
- Reimplement out_depth, lhs_depth and rhs_depth and basis from original compute code, with some FIXME comments for review where defaults are used.
- Note this is still buggy as python throws a malloc error when test_ft_fma is run. I have checked that out[idx] is not overrunning here which looks fine but most likely I'm passing different args than compute.
- Docstrings in for base roughpy_jax.
- Reverting indexing to S32 type so users are not forced into enabling JAX 64 bit features by default. For now this is implemented by copying values into a temp 64 bit array.
- Added TensorBasis size() method as in compute.

+++ roughpy_jax/tests/test_tensor_types.py
@@ -19,7 +19,7 @@ def test_tensor_basis_degree_begin(width, depth, expected):
     # Absolute check that the degree begin matches sample
     np.testing.assert_array_equal(
         tensor_basis.degree_begin,
-        np.array(expected, dtype=np.int64)
+        np.array(expected, dtype=np.int32)
     )
 
     # Defensive check that generated degree begin array has right size
